requirements.txt
flwr[simulation]==1.10.0
flwr-datasets==0.5.0
datasets==2.19.0
omegaconf==2.3.0
hydra-core==1.3.2
torch>=2.2.0
streamlit==1.47.1
matplotlib==3.10.5
pynvml==12.0.0